Refactor daily brief generation to utilize Claude API for content synthesis and remove AI synthesis service. Update email formatting and structure for improved readability and responsiveness.
